image-to-text/requirements.txt
pillow
tensorflow
numpy
fastapi
uvicorn
python-multipart
